@@ -1,14 +1,18 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-import { StoreProvider } from "easy-peasy";
-import { store } from "./redux/store";
-import { ThemeProvider } from "styled-components";
-import theme from "./theme";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+import reportWebVitals from './reportWebVitals';
+import { StoreProvider } from 'easy-peasy';
+import { store } from '@redux/store';
+import { ThemeProvider } from 'styled-components';
+import theme from './theme';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement,
+);
+
+root.render(
   <React.StrictMode>
     <StoreProvider store={store}>
       <ThemeProvider theme={theme}>
@@ -16,7 +20,6 @@ ReactDOM.render(
       </ThemeProvider>
     </StoreProvider>
   </React.StrictMode>,
-  document.getElementById("root")
 );
 
 // If you want to start measuring performance in your app, pass a function
