@@ -1,4 +1,4 @@
-import type { DeckDealerContext, DeckDealer } from "./types";
+import type { DeckDealerContext, DeckDealer } from './types';
 
 export default class DealerContext implements DeckDealerContext {
   constructor(private deckDealer: DeckDealer) {}
