@@ -1,4 +1,4 @@
-import DealerContext from "./DealerContext";
-import RandomDealer from "./dealers/RandomDealer";
+import DealerContext from './DealerContext';
+import RandomDealer from './dealers/RandomDealer';
 
 export { DealerContext, RandomDealer };
