@@ -1,4 +1,4 @@
-import DeckCreatorContext from "./DeckCreatorContext";
-import DefaultDeckCreator from "./creators/DefaultDeckCreator";
+import DeckCreatorContext from './DeckCreatorContext';
+import DefaultDeckCreator from './creators/DefaultDeckCreator';
 
 export { DeckCreatorContext, DefaultDeckCreator };
