@@ -1,8 +1,8 @@
-import type { HandEvaluator } from "../types";
-import type { Hand } from "../../../../../../global/types";
+import type { HandEvaluator } from '../types';
+import type { Hand } from '@global/types';
 
 interface Frequencies {
-  [key: string]: any;
+  [key: string]: number;
 }
 export default class FourOfAKindEvaluator implements HandEvaluator {
   constructor(private hand: Hand) {}
@@ -17,9 +17,8 @@ export default class FourOfAKindEvaluator implements HandEvaluator {
       frequencies[cardProperty] = ++frequencies[cardProperty] || 1;
     }
 
-    for (const [cardNumber, cardNumberFrequency] of Object.entries(
-      frequencies
-    )) {
+    // eslint-disable-next-line
+    for (const [_, cardNumberFrequency] of Object.entries(frequencies)) {
       if (cardNumberFrequency === 4) {
         isFourOfKind = true;
       }
