@@ -1,7 +1,7 @@
-import type { HandEvaluator } from "../types";
-import type { Hand } from "../../../../../../global/types";
-import PairEvaluator from "./PairEvaluator";
-import ThreeOfAKindEvaluator from "./ThreeOfAKindEvaluator";
+import type { HandEvaluator } from '../types';
+import type { Hand } from '@global/types';
+import PairEvaluator from './PairEvaluator';
+import ThreeOfAKindEvaluator from './ThreeOfAKindEvaluator';
 
 export default class FullHouseEvaluator implements HandEvaluator {
   constructor(private hand: Hand) {}
