@@ -1,5 +1,5 @@
-import type { HandEvaluator } from "../types";
-import type { Hand } from "../../../../../../global/types";
+import type { HandEvaluator } from '../types';
+import type { Hand } from '@global/types';
 
 export default class RoyalFlushEvaluator implements HandEvaluator {
   constructor(private hand: Hand) {}
