@@ -1,13 +1,13 @@
-import HandEvalutorsContext from "./HandEvalutorsContext";
-import RoyalFlushEvaluator from "./evaluators/RoyalFlushEvaluator";
-import StraighFlushEvaluator from "./evaluators/StraighFlushEvaluator";
-import FourOfAKindEvaluator from "./evaluators/FourOfAKindEvaluator";
-import FlushEvaluator from "./evaluators/FlushEvaluator";
-import StraightEvaluator from "./evaluators/StraightEvaluator";
-import ThreeOfAKindEvaluator from "./evaluators/ThreeOfAKindEvaluator";
-import PairEvaluator from "./evaluators/PairEvaluator";
-import TwoPairEvaluator from "./evaluators/TwoPairEvaluator";
-import FullHouseEvaluator from "./evaluators/FullHouseEvaluator";
+import HandEvalutorsContext from './HandEvalutorsContext';
+import RoyalFlushEvaluator from './evaluators/RoyalFlushEvaluator';
+import StraighFlushEvaluator from './evaluators/StraighFlushEvaluator';
+import FourOfAKindEvaluator from './evaluators/FourOfAKindEvaluator';
+import FlushEvaluator from './evaluators/FlushEvaluator';
+import StraightEvaluator from './evaluators/StraightEvaluator';
+import ThreeOfAKindEvaluator from './evaluators/ThreeOfAKindEvaluator';
+import PairEvaluator from './evaluators/PairEvaluator';
+import TwoPairEvaluator from './evaluators/TwoPairEvaluator';
+import FullHouseEvaluator from './evaluators/FullHouseEvaluator';
 
 export {
   HandEvalutorsContext,
