@@ -1,23 +1,23 @@
-import { HandEvalutorsContext, FourOfAKindEvaluator } from "../index";
+import { HandEvalutorsContext, FourOfAKindEvaluator } from '../index';
 import {
   mockLegitThreeOfAKind,
   mockLegitFourOfAKind,
-} from "./mocks/mock-hands";
+} from './mocks/mock-hands';
 
-describe("four-of-a-kind tests", () => {
-  test("is four of a kind", () => {
+describe('four-of-a-kind tests', () => {
+  test('is four of a kind', () => {
     expect(
       new HandEvalutorsContext(
-        new FourOfAKindEvaluator(mockLegitFourOfAKind)
-      ).evaluate()
+        new FourOfAKindEvaluator(mockLegitFourOfAKind),
+      ).evaluate(),
     ).toBe(true);
   });
 
-  test("is three of a kind", () => {
+  test('is three of a kind', () => {
     expect(
       new HandEvalutorsContext(
-        new FourOfAKindEvaluator(mockLegitThreeOfAKind)
-      ).evaluate()
+        new FourOfAKindEvaluator(mockLegitThreeOfAKind),
+      ).evaluate(),
     ).toBe(false);
   });
 });
