@@ -1,10 +1,10 @@
-import { HandEvalutorsContext, PairEvaluator } from "../index";
-import { mockLegitPair, mockLegitTwoPair } from "./mocks/mock-hands";
+import { HandEvalutorsContext, PairEvaluator } from '../index';
+import { mockLegitPair } from './mocks/mock-hands';
 
-describe("pair tests", () => {
-  test("is a pair", () => {
+describe('pair tests', () => {
+  test('is a pair', () => {
     expect(
-      new HandEvalutorsContext(new PairEvaluator(mockLegitPair)).evaluate()
+      new HandEvalutorsContext(new PairEvaluator(mockLegitPair)).evaluate(),
     ).toBe(true);
   });
 });
