@@ -1,41 +1,41 @@
-import { HandEvalutorsContext, StraighFlushEvaluator } from "../index";
+import { HandEvalutorsContext, StraighFlushEvaluator } from '../index';
 import {
   mockLegitStraightFlush,
   mockStraightFlushHandNotInSequence,
   mockStraightFlushHandNotOfSameSuit,
   mockHandLegitRoyalFlush,
-} from "./mocks/mock-hands";
+} from './mocks/mock-hands';
 
-describe("straight-flush tests", () => {
-  test("is a straight flush", () => {
+describe('straight-flush tests', () => {
+  test('is a straight flush', () => {
     expect(
       new HandEvalutorsContext(
-        new StraighFlushEvaluator(mockLegitStraightFlush)
-      ).evaluate()
+        new StraighFlushEvaluator(mockLegitStraightFlush),
+      ).evaluate(),
     ).toBe(true);
   });
 
-  test("is not of same suit", () => {
+  test('is not of same suit', () => {
     expect(
       new HandEvalutorsContext(
-        new StraighFlushEvaluator(mockStraightFlushHandNotOfSameSuit)
-      ).evaluate()
+        new StraighFlushEvaluator(mockStraightFlushHandNotOfSameSuit),
+      ).evaluate(),
     ).toBe(false);
   });
 
-  test("is not in sequence", () => {
+  test('is not in sequence', () => {
     expect(
       new HandEvalutorsContext(
-        new StraighFlushEvaluator(mockStraightFlushHandNotInSequence)
-      ).evaluate()
+        new StraighFlushEvaluator(mockStraightFlushHandNotInSequence),
+      ).evaluate(),
     ).toBe(false);
   });
 
-  test("is a royal flush", () => {
+  test('is a royal flush', () => {
     expect(
       new HandEvalutorsContext(
-        new StraighFlushEvaluator(mockHandLegitRoyalFlush)
-      ).evaluate()
+        new StraighFlushEvaluator(mockHandLegitRoyalFlush),
+      ).evaluate(),
     ).toBe(false);
   });
 });
