@@ -1,10 +1,10 @@
-import { HandEvalutorsContext, PairEvaluator } from "../index";
-import { mockLegitTwoPair } from "./mocks/mock-hands";
+import { HandEvalutorsContext, PairEvaluator } from '../index';
+import { mockLegitTwoPair } from './mocks/mock-hands';
 
-describe("two pair tests", () => {
-  test("is a two pair", () => {
+describe('two pair tests', () => {
+  test('is a two pair', () => {
     expect(
-      new HandEvalutorsContext(new PairEvaluator(mockLegitTwoPair)).evaluate()
+      new HandEvalutorsContext(new PairEvaluator(mockLegitTwoPair)).evaluate(),
     ).toBe(true);
   });
 });
