@@ -1,4 +1,4 @@
-import type { DeckShufflerContext, DeckShuffler } from "./types";
+import type { DeckShufflerContext, DeckShuffler } from './types';
 
 export default class ShufflerContext implements DeckShufflerContext {
   constructor(private deckShuffler: DeckShuffler) {}
