@@ -1,4 +1,4 @@
-import ShufflerContext from "./ShufflerContext";
-import RandomShuffler from "./algorithms/RandomShuffler";
+import ShufflerContext from './ShufflerContext';
+import RandomShuffler from './algorithms/RandomShuffler';
 
 export { ShufflerContext, RandomShuffler };
