@@ -1,4 +1,4 @@
-import { CardDeck } from "../../../../../global/types";
+import { CardDeck } from '@global/types';
 
 export interface DeckShuffler {
   shuffle: () => CardDeck;
