@@ -1,8 +1,8 @@
-import { action } from "easy-peasy";
-import type { AlertModel } from "../../../../redux/types";
+import { action } from 'easy-peasy';
+import type { AlertModel } from '@redux/types';
 
 const alertsModel: AlertModel = {
-  alertConfig: { isVisible: false, message: "" },
+  alertConfig: { isVisible: false, message: '' },
   configureAlert: action((state, payload) => {
     state.alertConfig = payload;
   }),
