@@ -1,13 +1,17 @@
-import styled, { DefaultTheme } from "styled-components";
-import { animated } from 'react-spring';
+import styled, { DefaultTheme } from 'styled-components';
+import { animated, SpringValue } from 'react-spring';
+
+type AnimatedStyle = {
+  opacity: SpringValue<number>;
+};
 
 interface StyledPokerPlayerProps {
   theme: DefaultTheme;
-  style: any;
+  style: AnimatedStyle;
 }
 
 const StyledPokerPlayer = styled(animated.div)<StyledPokerPlayerProps>`
-  display: "flex";   
+  display: 'flex';
   flex-direction: column;
   flex-wrap: wrap;
   justify-content: flex-start;
