@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import styled from 'styled-components';
 
 const StyledPokerTable = styled.div`
   display: flex;
@@ -7,7 +7,7 @@ const StyledPokerTable = styled.div`
   justify-content: flex-start;
   background-color: ${(props) => props.theme.colors.primary};
   border-radius: 10px;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2)
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2);
 `;
 
 export { StyledPokerTable };
