@@ -1,5 +1,5 @@
-import styled from "styled-components";
-import { Container } from "react-bootstrap";
+import styled from 'styled-components';
+import { Container } from 'react-bootstrap';
 
 const StyledScreenTemplate = styled(Container)`
   padding: 5%;
