@@ -1,6 +1,7 @@
-import { memo, FC } from "react";
-import { ScreenTemplate } from "../components/templates/ScreenTemplate";
-import { PokerGame } from "../components/containers/PokerGame";
+import { memo, FC } from 'react';
+
+import { ScreenTemplate } from '@components/templates/ScreenTemplate';
+import { PokerGame } from '@components/containers/PokerGame';
 
 const GamePage: FC = () => {
   return (
