@@ -1,1 +1,1 @@
-export { default as GamePage } from "./GamePage";
+export { default as GamePage } from './GamePage';
