@@ -1,6 +1,6 @@
-export { models } from "./models";
+export { models } from './models';
 export {
   useStoreActions,
   useStoreDispatch,
   useStoreState,
-} from "./typed-hooks";
+} from './typed-hooks';
