@@ -1,17 +1,16 @@
-
 const theme = {
   colors: {
-    primary: "#0C361F",
+    primary: '#0C361F',
     secondary: '#800020',
-    accent: "#FFD700",
-    dark: "#000000",
-    success: "#28a745",
-    info: "#17a2b8",
-    warning: "#ffc107",
-    danger: "#dc3545",
-    light: "#f8f9fa"
+    accent: '#FFD700',
+    dark: '#000000',
+    success: '#28a745',
+    info: '#17a2b8',
+    warning: '#ffc107',
+    danger: '#dc3545',
+    light: '#f8f9fa',
   },
-  fonts: ["Arial", "sans-serif"]
+  fonts: ['Arial', 'sans-serif'],
 };
 
 export default theme;
